Revert "renamed config.h to resource.h"
This reverts commit 9c96f584bab729a87c65d7cd7f59b6d9e1d83f1e.
