+++ Config.h
@@ -1,0 +1,6 @@
+#ifndef CONFIG_H
+#define CONFIG_H
+
+const int MAPSIZE = 30;
+
+#endif
