--- Resource.h
@@ -1,7 +1,0 @@
-#ifndef RESOURCE_H
-#define RESOURCE_H
-
-const int MAPSIZE = 30;
-int turnCounter = 0;
-
-#endif
